Single slot/requirements.txt
ultralytics==8.4.26
opencv-python==4.13.0.92
pyserial==3.5
firebase==4.0.1
torch==2.11.0
torchvision==0.26.0
numpy==2.4.3